Adjust `kaff4 verify` command to simplify code
This is a NO-OP refactor
